% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{run app}
\usage{
run_app()
}
\description{
Run the ImproveYourSelfesteem app
}
\references{
De Neef, M. (2022). Negatief zelfbeeld (Revised). Boom.
De Neef, M. (z.d.). https://www.negatiefzelfbeeldbehandelen.nl/
}
